requirements.txt
streamlit
torch
torchvision
scikit-learn
matplotlib
seaborn
pillow
